+++ src/movies/MovieModel.cpp
@@ -250,14 +250,12 @@ QVector<Movie*> MovieModel::movies()
     return m_movies;
 }
 
-/**
- * @brief Checks if there are new movies (movies where infoLoaded is false)
- * @return True if there are new movies
- */
-long MovieModel::countNewMovies()
+/// @brief Checks if there are new movies (movies where infoLoaded is false)
+/// @return True if there are new movies
+int MovieModel::countNewMovies()
 {
     const auto checkInfoLoaded = [](const Movie* movie) { return !movie->controller()->infoLoaded(); };
-    return std::count_if(m_movies.cbegin(), m_movies.cend(), checkInfoLoaded);
+    return static_cast<int>(std::count_if(m_movies.cbegin(), m_movies.cend(), checkInfoLoaded));
 }
 
 int MovieModel::mediaStatusToColumn(MediaStatusColumn column)
